requirements.txt
Django>=4.2,<5.0
openai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0

